requirements.txt
pygame == 2.5.2
pyyaml == 6.0.1
